Artificial Intelligence (AI)/Projects/Agentic AI/PhiData/1. Financial Agent with PhiData/requirements.txt
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
psutil